pyproject.toml
[project]
name = "langchain-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.0.0",
]
